##Fixed error??
Updated the code for doing the year by year analysis. What was going wrong was that for tenure group 10 years and over etc there was not enough ppl per year so no reg possible which stopped the function. We wrote in a try() routine to skip.

Not sure why this is an issue now and not in the past -- follow up.